+++ apps/src/lib/client/tx.rs
@@ -500,177 +500,13 @@ impl masp::ShieldedUtils for CLIShieldedUtils {
     }
 }
 
-
-
 pub async fn submit_transfer<C: Client + namada::ledger::queries::Client + Sync, V: WalletUtils, U: ShieldedUtils<C = C>, P>(
     client: &C,
     wallet: &mut Wallet<P>,
     shielded: &mut ShieldedContext<U>,
     args: args::TxTransfer,
 ) {
-    let transfer_source = args.source;
-    let source = transfer_source.effective_address();
-    let transfer_target = args.target.clone();
-    let target = transfer_target.effective_address();
-    // Check that the source address exists on chain
-    let source_exists =
-        rpc::known_address::<C>(client, &source).await;
-    if !source_exists {
-        eprintln!("The source address {} doesn't exist on chain.", source);
-        if !args.tx.force {
-            safe_exit(1)
-        }
-    }
-    // Check that the target address exists on chain
-    let target_exists =
-        rpc::known_address::<C>(client, &target).await;
-    if !target_exists {
-        eprintln!("The target address {} doesn't exist on chain.", target);
-        if !args.tx.force {
-            safe_exit(1)
-        }
-    }
-    let token = &args.token;
-    // Check that the token address exists on chain
-    let token_exists =
-        rpc::known_address::<C>(client, &token)
-            .await;
-    if !token_exists {
-        eprintln!(
-            "The token address {} doesn't exist on chain.",
-            token
-        );
-        if !args.tx.force {
-            safe_exit(1)
-        }
-    }
-    // Check source balance
-    let (sub_prefix, balance_key) = match &args.sub_prefix {
-        Some(sub_prefix) => {
-            let sub_prefix = storage::Key::parse(sub_prefix).unwrap();
-            let prefix = token::multitoken_balance_prefix(
-                &token,
-                &sub_prefix,
-            );
-            (
-                Some(sub_prefix),
-                token::multitoken_balance_key(&prefix, &source),
-            )
-        }
-        None => (None, token::balance_key(&token, &source)),
-    };
-    match rpc::query_storage_value::<C, token::Amount>(&client, &balance_key).await
-    {
-        Some(balance) => {
-            if balance < args.amount {
-                eprintln!(
-                    "The balance of the source {} of token {} is lower than \
-                     the amount to be transferred. Amount to transfer is {} \
-                     and the balance is {}.",
-                    source, token, args.amount, balance
-                );
-                if !args.tx.force {
-                    safe_exit(1)
-                }
-            }
-        }
-        None => {
-            eprintln!(
-                "No balance found for the source {} of token {}",
-                source, token
-            );
-            if !args.tx.force {
-                safe_exit(1)
-            }
-        }
-    };
-
-    let tx_code = args.tx_code_path;
-    let masp_addr = masp();
-    // For MASP sources, use a special sentinel key recognized by VPs as default
-    // signer. Also, if the transaction is shielded, redact the amount and token
-    // types by setting the transparent value to 0 and token type to a constant.
-    // This has no side-effect because transaction is to self.
-    let (default_signer, amount, token) =
-        if source == masp_addr && target == masp_addr {
-            // TODO Refactor me, we shouldn't rely on any specific token here.
-            (
-                TxSigningKey::SecretKey(masp_tx_key()),
-                0.into(),
-                args.native_token.clone(),
-            )
-        } else if source == masp_addr {
-            (
-                TxSigningKey::SecretKey(masp_tx_key()),
-                args.amount,
-                token.clone(),
-            )
-        } else {
-            (
-                TxSigningKey::WalletAddress(source.clone()),
-                args.amount,
-                token.clone(),
-            )
-        };
-    // If our chosen signer is the MASP sentinel key, then our shielded inputs
-    // will need to cover the gas fees.
-    let chosen_signer = tx_signer::<C, V, P>(client, wallet, &args.tx, default_signer.clone())
-        .await
-        .ref_to();
-    let shielded_gas = masp_tx_key().ref_to() == chosen_signer;
-    // Determine whether to pin this transaction to a storage key
-    let key = match &args.target {
-        TransferTarget::PaymentAddress(pa) if pa.is_pinned() => Some(pa.hash()),
-        _ => None,
-    };
-
-    let stx_result =
-        shielded.gen_shielded_transfer(
-            client,
-            transfer_source,
-            transfer_target,
-            args.amount,
-            args.token,
-            args.tx.fee_amount,
-            args.tx.fee_token.clone(),
-            shielded_gas,
-        )
-        .await;
-    let shielded = match stx_result {
-        Ok(stx) => stx.map(|x| x.0),
-        Err(builder::Error::ChangeIsNegative(_)) => {
-            eprintln!(
-                "The balance of the source {} is lower than the \
-                 amount to be transferred and fees. Amount to \
-                 transfer is {} {} and fees are {} {}.",
-                source,
-                args.amount,
-                token,
-                args.tx.fee_amount,
-                &args.tx.fee_token,
-            );
-            safe_exit(1)
-        }
-        Err(err) => panic!("{}", err),
-    };
-
-    let transfer = token::Transfer {
-        source: source.clone(),
-        target,
-        token,
-        sub_prefix,
-        amount,
-        key,
-        shielded,
-    };
-    tracing::debug!("Transfer data {:?}", transfer);
-    let data = transfer
-        .try_to_vec()
-        .expect("Encoding tx data shouldn't fail");
-
-    let tx = Tx::new(tx_code, Some(data));
-    let signing_address = TxSigningKey::WalletAddress(source);
-    process_tx::<C, V, P>(client, wallet, &args.tx, tx, signing_address).await;
+    namada::ledger::tx::submit_transfer::<C, V, U, P>(client, wallet, shielded, args).await;
 }
 
 pub async fn submit_ibc_transfer<C: Client + namada::ledger::queries::Client + Sync, U: WalletUtils, P>(
@@ -678,113 +514,7 @@ pub async fn submit_ibc_transfer<C: Client + namada::ledger::queries::Client + S
     wallet: &mut Wallet<P>,
     args: args::TxIbcTransfer,
 ) {
-    let source = args.source.clone();
-    // Check that the source address exists on chain
-    let source_exists =
-        rpc::known_address::<C>(client, &source).await;
-    if !source_exists {
-        eprintln!("The source address {} doesn't exist on chain.", source);
-        if !args.tx.force {
-            safe_exit(1)
-        }
-    }
-
-    // We cannot check the receiver
-
-    let token = args.token;
-    // Check that the token address exists on chain
-    let token_exists =
-        rpc::known_address::<C>(client, &token).await;
-    if !token_exists {
-        eprintln!("The token address {} doesn't exist on chain.", token);
-        if !args.tx.force {
-            safe_exit(1)
-        }
-    }
-    // Check source balance
-    let (sub_prefix, balance_key) = match args.sub_prefix {
-        Some(sub_prefix) => {
-            let sub_prefix = storage::Key::parse(sub_prefix).unwrap();
-            let prefix = token::multitoken_balance_prefix(&token, &sub_prefix);
-            (
-                Some(sub_prefix),
-                token::multitoken_balance_key(&prefix, &source),
-            )
-        }
-        None => (None, token::balance_key(&token, &source)),
-    };
-    match rpc::query_storage_value::<C, token::Amount>(&client, &balance_key).await
-    {
-        Some(balance) => {
-            if balance < args.amount {
-                eprintln!(
-                    "The balance of the source {} of token {} is lower than \
-                     the amount to be transferred. Amount to transfer is {} \
-                     and the balance is {}.",
-                    source, token, args.amount, balance
-                );
-                if !args.tx.force {
-                    safe_exit(1)
-                }
-            }
-        }
-        None => {
-            eprintln!(
-                "No balance found for the source {} of token {}",
-                source, token
-            );
-            if !args.tx.force {
-                safe_exit(1)
-            }
-        }
-    }
-    let tx_code = args.tx_code_path;
-
-    let denom = match sub_prefix {
-        // To parse IbcToken address, remove the address prefix
-        Some(sp) => sp.to_string().replace(RESERVED_ADDRESS_PREFIX, ""),
-        None => token.to_string(),
-    };
-    let token = Some(Coin {
-        denom,
-        amount: args.amount.to_string(),
-    });
-
-    // this height should be that of the destination chain, not this chain
-    let timeout_height = match args.timeout_height {
-        Some(h) => IbcHeight::new(0, h),
-        None => IbcHeight::zero(),
-    };
-
-    let now: namada::tendermint::Time = DateTimeUtc::now().try_into().unwrap();
-    let now: IbcTimestamp = now.into();
-    let timeout_timestamp = if let Some(offset) = args.timeout_sec_offset {
-        (now + Duration::new(offset, 0)).unwrap()
-    } else if timeout_height.is_zero() {
-        // we cannot set 0 to both the height and the timestamp
-        (now + Duration::new(3600, 0)).unwrap()
-    } else {
-        IbcTimestamp::none()
-    };
-
-    let msg = MsgTransfer {
-        source_port: args.port_id,
-        source_channel: args.channel_id,
-        token,
-        sender: Signer::new(source.to_string()),
-        receiver: Signer::new(args.receiver),
-        timeout_height,
-        timeout_timestamp,
-    };
-    tracing::debug!("IBC transfer message {:?}", msg);
-    let any_msg = msg.to_any();
-    let mut data = vec![];
-    prost::Message::encode(&any_msg, &mut data)
-        .expect("Encoding tx data shouldn't fail");
-
-    let tx = Tx::new(tx_code, Some(data));
-    process_tx::<C, U, P>(client, wallet, &args.tx, tx, TxSigningKey::WalletAddress(args.source))
-        .await;
+    namada::ledger::tx::submit_ibc_transfer::<C, U, P>(client, wallet, args).await;
 }
 
 pub async fn submit_init_proposal<C: Client + namada::ledger::queries::Client + Sync, U: WalletUtils>(
@@ -1119,27 +849,7 @@ async fn is_safe_voting_window<C: Client + namada::ledger::queries::Client + Syn
     proposal_id: u64,
     proposal_start_epoch: Epoch,
 ) -> bool {
-    let current_epoch = rpc::query_epoch(client).await;
-
-    let proposal_end_epoch_key =
-        gov_storage::get_voting_end_epoch_key(proposal_id);
-    let proposal_end_epoch =
-        rpc::query_storage_value::<C, Epoch>(client, &proposal_end_epoch_key)
-            .await;
-
-    match proposal_end_epoch {
-        Some(proposal_end_epoch) => {
-            !namada::ledger::native_vp::governance::utils::is_valid_validator_voting_period(
-                current_epoch,
-                proposal_start_epoch,
-                proposal_end_epoch,
-            )
-        }
-        None => {
-            eprintln!("Proposal end epoch is not in the storage.");
-            safe_exit(1)
-        }
-    }
+    namada::ledger::tx::is_safe_voting_window(client, proposal_id, proposal_start_epoch).await
 }
 
 /// Removes validators whose vote corresponds to that of the delegator (needless
